generalize metadata query service
