+++ api/routers/filer/routes/root.py
@@ -3,11 +3,11 @@ from typing import Annotated
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from api.common.exceptions import RESPONSES
+from api.common.services.metadata_query import MetadataQueryService
 from api.models.response_model_properties import RequestDataModel
 from api.models.base_response_models import BaseResponseModel
 
 from api.routers.filer.common.constants import ROUTE_NAME, ROUTE_TAGS, ROUTE_PREFIX
-from api.routers.filer.common.services import MetadataQueryService
 from api.routers.filer.dependencies.parameters import ROUTE_SESSION_MANAGER
 
 from api.routers.filer.routes.track import router as TrackRouter
